Improve refresh button
